backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Web Scraping
beautifulsoup4==4.12.3
trafilatura==1.7.0
requests==2.31.0
urllib3==2.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.4

# PowerPoint Generation
python-pptx==0.6.23

# OpenAI
openai==1.12.0

# Image Processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2

# Development
ipython==8.21.0
django-debug-toolbar==4.3.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

